Only upload changed files during deploy (closes #6)
Git does not record timestamps for files, as there is no central notion of time and date (distributed VCS). This means that, when a repo is cloned, all checked out entries' modification times are the current time. By extension, files synced to S3 have the same date as the clone in the build when they were deployed: the client uploads if any of the following are true:

 - The local file is a different size.
 - The last modified time of the local file is more recent.
    - Unless `--size-only` is passed.
 - The local file does not exist remotely.

We fix this by setting the modified time of each file to the time of the last commit that changed that file. This means only files changed since the last deploy will be uploaded. However, if a file is changed, then changed back, it will still be re-uploaded, even though the files are the same. This could be fixed by looking at what has changed since the last tag, however is is not a common enough use case to warrant the complexity. It can always be added later if the need arises.